+++ src/app/auth/login/login-form.tsx
@@ -1,0 +1,65 @@
+'use client';
+
+import Link from 'next/link';
+import { Controller } from 'react-hook-form';
+
+import { Button } from '@/components/ui/button';
+import { Field, FieldGroup, FieldLabel, FieldError } from '@/components/ui/field';
+import { Input } from '@/components/ui/input';
+
+import useLoginForm from './use-login-form';
+
+export function LoginForm() {
+  const { form, onSubmit } = useLoginForm();
+
+  return (
+    <form onSubmit={form.handleSubmit(onSubmit)} className="'flex gap-6' flex-col">
+      <FieldGroup>
+        <Controller
+          name="email"
+          control={form.control}
+          render={({ field, fieldState }) => (
+            <Field data-invalid={fieldState.invalid}>
+              <FieldLabel htmlFor="email">Email</FieldLabel>
+              <Input
+                {...field}
+                id="email"
+                aria-invalid={fieldState.invalid}
+                placeholder="you@example.com"
+                autoComplete="email"
+              />
+              {fieldState.invalid && <FieldError errors={[fieldState.error]} />}
+            </Field>
+          )}
+        />
+        <Controller
+          name="password"
+          control={form.control}
+          render={({ field, fieldState }) => (
+            <Field data-invalid={fieldState.invalid}>
+              <div className="flex items-center">
+                <FieldLabel htmlFor="password">Password</FieldLabel>
+                <Link href="/auth/forgot-password" className="ml-auto text-sm underline-offset-4 hover:underline">
+                  Forgot your password?
+                </Link>
+              </div>
+              <Input
+                {...field}
+                id="password"
+                aria-invalid={fieldState.invalid}
+                placeholder="••••••••"
+                autoComplete="off"
+              />
+              {fieldState.invalid && <FieldError errors={[fieldState.error]} />}
+            </Field>
+          )}
+        />
+        <Field>
+          <Button type="submit" className="cursor-pointer">
+            Login
+          </Button>
+        </Field>
+      </FieldGroup>
+    </form>
+  );
+}